Split out base functionality into a more generic class
